skal utlede miljøvariable via yaml-fil. SKal laste opp mindre bloat i dockercontainer
