+++ Dockerfile
@@ -1,7 +1,13 @@
 FROM navikt/node-express:18
 
+WORKDIR /var/server
+
+COPY dist ./dist
+COPY index.html ./index.html
+COPY ./src/backend/server.js ./server.js
+
 ADD ./ /var/server/
 
 EXPOSE 8080
 
-CMD ["node", "src/backend/server.js"]
+CMD ["node", "server.js"]
