+++ nais-prod.yaml
@@ -29,6 +29,8 @@ spec:
   env:
     - name: APP_VERSION
       value: '{{version}}'
+    - name: ENV
+      value: production
   resources:
     limits:
       memory: 512Mi
