+++ src/backend/server.js
@@ -1,7 +1,6 @@
-import path from 'path';
+import dekorator from '@navikt/nav-dekoratoren-moduler/ssr/index.js';
 import express from 'express';
-import indexHandler from './dekorator.js';
-import appEnv from './environment.js';
+import path from 'path';
 
 const app = express();
 const port = 8080;
@@ -9,14 +8,26 @@ const port = 8080;
 const BASE_PATH = '/familie/alene-med-barn/veiviser';
 const byggmappeFrontend = path.join(process.cwd(), 'dist');
 
+const indexHandler = (_, res) => {
+  dekorator
+    .injectDecoratorServerSide({
+      env: process.env.ENV === 'production' ? 'prod' : 'dev',
+      filePath: `${path.join(process.cwd(), 'dist')}/index.html`,
+    })
+    .then((html) => {
+      res.send(html);
+    })
+    .catch((e) => {
+      console.log(e);
+      const error = `En feil oppstod. Klikk <a href='https://www.nav.no'>her</a> for å gå tilbake til nav.no. Kontakt kundestøtte hvis problemet vedvarer.`;
+      res.status(500).send(error);
+    });
+};
+
 app.set('views', byggmappeFrontend);
 
 app.get(`${BASE_PATH}/status`, (_req, res) => {
   res.status(200).end();
-});
-
-app.get(`${BASE_PATH}/env`, (_req, res) => {
-  res.status(200).send(appEnv).end();
 });
 
 app.use(
